feat(publish): move Default Share Image card to Domains & Publishing settings
Master's PR #1718 split publishing settings into a dedicated
"Domains & Publishing" page. The drive-level default OG image belongs
with the other publishing controls, not in General — move the card (and
its state/handler) to settings/domains and drop it from settings/general.
